backend/requirements.txt
flask
pywebview
CORS
flask_cors
psutil
routes
pyinstaller